feat: add logic for all lists truncation
