+++ dynamiq/types/document.py
@@ -1,10 +1,10 @@
 import enum
 import uuid
-from typing import Any, Callable, TypedDict
+from typing import Any, Callable
 
 from pydantic import BaseModel, Field
 
-from dynamiq.utils.utils import TRUNCATE_LIMIT
+from dynamiq.utils.utils import TRUNCATE_EMBEDDINGS_LIMIT
 
 
 class Document(BaseModel):
@@ -23,7 +23,7 @@ class Document(BaseModel):
     embedding: list | None = None
     score: float | None = None
 
-    def to_dict(self, for_tracing: bool = False, truncate_limit: int = TRUNCATE_LIMIT, **kwargs) -> dict:
+    def to_dict(self, for_tracing: bool = False, truncate_limit: int = TRUNCATE_EMBEDDINGS_LIMIT, **kwargs) -> dict:
         """Convert the Document object to a dictionary.
 
         Returns:
@@ -44,22 +44,3 @@ class DocumentCreationMode(str, enum.Enum):
     ONE_DOC_PER_FILE = "one-doc-per-file"
     ONE_DOC_PER_PAGE = "one-doc-per-page"
     ONE_DOC_PER_ELEMENT = "one-doc-per-element"
-
-
-class TextEmbeddingOutputDict(TypedDict):
-    embedding: list[float]
-    query: str
-
-
-class TextEmbeddingOutput(dict):
-    """Dict-like output for text embedders with tracing-aware to_dict()."""
-
-    query: str
-    embedding: list[float]
-
-    def to_dict(self, for_tracing: bool = False, truncate_limit: int = TRUNCATE_LIMIT, **kwargs) -> dict:
-        data = dict(self)
-        if for_tracing and isinstance(data.get("embedding"), list):
-            if len(data["embedding"]) > truncate_limit:
-                data["embedding"] = data["embedding"][:truncate_limit]
-        return data
